feat(estoque): adiciona campo de quantidade em Lote e regra de movimentação automática
- Adicionado o campo 'quantidade' na tabela Lote (salvo na menor unidade de medida).
- Implementada lógica para preenchimento de lotes iniciais vazios pós-cadastro do insumo.
- Criada validação para comparar a soma dos lotes com o estoque atual, gerando movimentação de entrada automática quando a quantidade total for ultrapassada.

+++ K-Gest/BancoDados/MovimentacaoEstoque.cs
@@ -7,7 +7,7 @@ namespace K_Gest.BancoDados
     {
         public int? idEstoque;
         public string tipoEs; // "E" para Entrada, "S" para Saída
-        public int qtdMoviment;
+        public decimal qtdMoviment;
         public string motivo;
         public int idInsumo;
